MODCITEM-1: Create mod-circulation-item. improvements part #2
