+++ src/main/java/org/folio/circulation/item/domain/mapper/CirculationItemMapper.java
@@ -5,15 +5,11 @@ import org.folio.rs.domain.dto.CirculationItem;
 import org.mapstruct.Mapper;
 import org.mapstruct.NullValueCheckStrategy;
 
-import java.util.List;
-
 @Mapper(componentModel = "spring", nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS)
 public interface CirculationItemMapper {
 
   CirculationItem mapEntityToDto(Item circulationItem);
 
   Item mapDtoToEntity(CirculationItem circulationItem);
-
-  List<CirculationItem> mapEntitiesToDtos(Iterable<Item> itemList);
 
 }
